Complete Phase 1: Replace Game interaction methods with interact:: calls
- Move interaction to a separate class.

This completes the Phase 1 refactoring by moving ~527 lines of interaction code from game.cpp to interact.cpp, reducing game.cpp complexity by ~12%.

All Game class methods now delegate to the interact:: namespace functions, maintaining the same API while improving code organization.
