+++ src/client/CMakeLists.txt
@@ -54,6 +54,8 @@ set(client_SRCS
 	${CMAKE_CURRENT_SOURCE_DIR}/game.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/gameui.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/game_formspec.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/interact.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/soundmaker.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/guiscalingfilter.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/hud.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/imagefilters.cpp
